requirements.txt
flask==2.0.1
python-dotenv==0.19.0
opencv-python==4.5.3.56
pillow==8.4.0
numpy==1.21.2